Cleaned EventTriggerer from the event timers and renamed to StoryManager and refactored the required stuff
