+++ Nitrox.Test/Server/GameLogic/StoryManagerTest.cs
@@ -1,0 +1,11 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using NitroxModel.Helper;
+
+namespace NitroxServer.GameLogic;
+
+[TestClass]
+public class StoryManagerTest
+{
+    private readonly string seed = StringHelper.GenerateRandomString(10);
+
+}
