+++ NitroxServer/Communication/Packets/Processors/BedEnterProcessor.cs
@@ -1,4 +1,4 @@
-﻿using NitroxModel.Packets;
+using NitroxModel.Packets;
 using NitroxServer.Communication.Packets.Processors.Abstract;
 using NitroxServer.GameLogic;
 
@@ -6,16 +6,17 @@ namespace NitroxServer.Communication.Packets.Processors
 {
     public class BedEnterProcessor : AuthenticatedPacketProcessor<BedEnter>
     {
-        private readonly EventTriggerer eventTriggerer;
+        private readonly StoryManager storyManager;
 
-        public BedEnterProcessor(EventTriggerer eventTriggerer)
+        public BedEnterProcessor(StoryManager storyManager)
         {
-            this.eventTriggerer = eventTriggerer;
+            this.storyManager = storyManager;
         }
 
         public override void Process(BedEnter packet, Player player)
         {
-            eventTriggerer.ChangeTime(EventTriggerer.TimeModification.SKIP);
+            // TODO: Needs repair since the new time implementation only relies on server-side time.
+            // storyManager.ChangeTime(StoryManager.TimeModification.SKIP);
         }
     }
 }
