+++ NitroxServer/ConsoleCommands/AuroraCommand.cs
@@ -1,21 +1,21 @@
-﻿using NitroxModel.DataStructures.GameLogic;
+using NitroxModel.DataStructures.GameLogic;
 using NitroxServer.ConsoleCommands.Abstract;
 using NitroxServer.ConsoleCommands.Abstract.Type;
 using NitroxServer.GameLogic;
 
 namespace NitroxServer.ConsoleCommands;
 
-// TODO: When we have the command system, we'll need to move this stuff to the new system
+// TODO: When we make the new command system, move this stuff to it
 public class AuroraCommand : Command
 {
-    private readonly EventTriggerer eventTriggerer;
+    private readonly StoryManager storyManager;
 
     // We shouldn't let the server use this command because it needs some stuff to happen client-side like goals
-    public AuroraCommand(EventTriggerer eventTriggerer) : base("aurora", Perms.ADMIN, PermsFlag.NO_CONSOLE, "Manage Aurora's state")
+    public AuroraCommand(StoryManager storyManager) : base("aurora", Perms.ADMIN, PermsFlag.NO_CONSOLE, "Manage Aurora's state")
     {
         AddParameter(new TypeString("countdown/restore/explode", true, "Which action to apply to Aurora"));
 
-        this.eventTriggerer = eventTriggerer;
+        this.storyManager = storyManager;
     }
 
     protected override void Execute(CallArgs args)
@@ -25,13 +25,13 @@ public class AuroraCommand : Command
         switch (action.ToLower())
         {
             case "countdown":
-                eventTriggerer.ExplodeAurora(true);
+                storyManager.ExplodeAurora(true);
                 break;
             case "restore":
-                eventTriggerer.RestoreAurora();
+                storyManager.RestoreAurora();
                 break;
             case "explode":
-                eventTriggerer.ExplodeAurora(false);
+                storyManager.ExplodeAurora(false);
                 break;
             default:
                 // Same message as in the abstract class, in method TryExecute
