+++ NitroxServer/Serialization/World/PersistedWorldData.cs
@@ -34,7 +34,7 @@ namespace NitroxServer.Serialization.World
                     ParsedBatchCells = world.BatchEntitySpawner.SerializableParsedBatches,
                     VehicleData = VehicleData.From(world.VehicleManager.GetVehicles()),
                     InventoryData = InventoryData.From(world.InventoryManager.GetAllStorageSlotItems(), world.InventoryManager.GetAllModules()),
-                    GameData = GameData.From(world.GameData.PDAState, world.GameData.StoryGoals, world.ScheduleKeeper, world.EventTriggerer),
+                    GameData = GameData.From(world.GameData.PDAState, world.GameData.StoryGoals, world.ScheduleKeeper, world.StoryManager),
                     Seed = world.Seed
                 }
             };
